Moved study materials types into separate file
