+++ src/components/SubjectMeanings/UserMeaningChips.tsx
@@ -6,7 +6,7 @@ import { generateUUID } from "../../utils";
 import { useStudyMaterialsBySubjIDs } from "../../hooks/useStudyMaterialsBySubjIDs";
 import { useStudyMaterialsChange } from "../../hooks/useStudyMaterialsChange";
 import { Subject } from "../../types/Subject";
-import { StudyMaterialDataResponse } from "../../types/MiscTypes";
+import { StudyMaterialDataResponse } from "../../types/StudyMaterial";
 import AlertModal from "../AlertModal";
 import { Chip } from "../../styles/BaseStyledComponents";
 import styled from "styled-components";
