+++ src/components/UserNote/Note.tsx
@@ -3,8 +3,9 @@ import { IonIcon } from "@ionic/react";
 import { capitalizeWord } from "../../services/MiscService";
 import { displayToast } from "../Toast/Toast.service";
 import { useStudyMaterialsChange } from "../../hooks/useStudyMaterialsChange";
-import { StudyMaterialDataResponse, UserNoteType } from "../../types/MiscTypes";
+import { UserNoteType } from "../../types/MiscTypes";
 import { Subject } from "../../types/Subject";
+import { StudyMaterialDataResponse } from "../../types/StudyMaterial";
 import AlertModal from "../AlertModal";
 import NoteIcon from "../../images/note.svg";
 import PencilIcon from "../../images/pencil.svg";
