+++ src/components/UserNote/UserNote.tsx
@@ -3,7 +3,8 @@ import { IonSkeletonText } from "@ionic/react";
 import { capitalizeWord } from "../../services/MiscService";
 import { useStudyMaterialsBySubjIDs } from "../../hooks/useStudyMaterialsBySubjIDs";
 import { Subject } from "../../types/Subject";
-import { StudyMaterialDataResponse, UserNoteType } from "../../types/MiscTypes";
+import { UserNoteType } from "../../types/MiscTypes";
+import { StudyMaterialDataResponse } from "../../types/StudyMaterial";
 import Note from "./Note";
 import { Chip } from "../../styles/BaseStyledComponents";
 import styled from "styled-components";
